Product repository, service, resource, GET ok

+++ backend/src/main/java/com/helion/dscatalog/entities/Product.java
@@ -26,7 +26,7 @@ public class Product implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	
 	@Column(columnDefinition = "TEXT")
@@ -52,7 +52,7 @@ public class Product implements Serializable{
 	
 	public Product() {}
 
-	public Product(Integer id, String name, String description, Instant date, Double price, String imgUrl) {
+	public Product(Long id, String name, String description, Instant date, Double price, String imgUrl) {
 		
 		this.id = id;
 		this.name = name;
@@ -63,12 +63,12 @@ public class Product implements Serializable{
 	}
 
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
